📝 Rename: 프리셋 관련 상수 대소문자 이슈로 파일명 변경
